refactor(inference): address review feedback for direct_json format
- Import DirectJsonParser and DirectJsonFormat from direct_json package root
- Remove inline import of DirectJsonParser in tests
- Raise A2uiCatalogError instead of ValueError when no supported catalogs configured
- Add test verifying A2uiCatalogError for missing supported catalogs
