+++ agent_sdks/python/a2ui_agent/tests/conformance/test_conformance.py
@@ -18,9 +18,8 @@ import pytest
 
 from a2ui.basic_catalog import BasicCatalog
 from a2ui.schema.catalog import A2uiCatalog
-from a2ui.inference_formats.direct_json.streaming import DirectJsonStreamParser
+from a2ui.inference_formats.direct_json import DirectJsonFormat, DirectJsonStreamParser
 from a2ui.validation.validator import A2uiValidator
-from a2ui.inference_formats.direct_json import DirectJsonFormat
 from a2ui.schema.catalog import CatalogConfig
 from a2ui.schema.common_modifiers import remove_strict_validation
 from a2ui.schema.constants import VERSION_0_8, VERSION_0_9
